+++ app/src/main/java/com/kirilpetriv/showcase/commonUi/ErrorView.kt
@@ -1,4 +1,4 @@
-package com.kirilpetriv.showcase.presentation.basecomposables
+package com.kirilpetriv.showcase.commonUi
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
