+++ app/src/main/java/com/kirilpetriv/showcase/di/ShowcaseModule.kt
@@ -3,8 +3,8 @@ package com.kirilpetriv.showcase.di
 import com.kirilpetriv.showcase.core.ArtworkRepository
 import com.kirilpetriv.showcase.data.ArtworkRepositoryImpl
 import com.kirilpetriv.showcase.network.ArtworkService
-import com.kirilpetriv.showcase.presentation.ArtworksDetailViewModel
-import com.kirilpetriv.showcase.presentation.ArtworksListViewModel
+import com.kirilpetriv.showcase.feature.artwork.ArtworksDetailViewModel
+import com.kirilpetriv.showcase.feature.artwork.ArtworksListViewModel
 import org.koin.core.module.dsl.viewModel
 import org.koin.dsl.module
