+++ app/src/main/java/com/kirilpetriv/showcase/feature/artwork/ArtworkExt.kt
@@ -1,4 +1,4 @@
-package com.kirilpetriv.showcase.presentation
+package com.kirilpetriv.showcase.feature.artwork
 
 import android.text.Html
 import android.text.Spanned
