+++ app/src/main/java/com/kirilpetriv/showcase/feature/artwork/ArtworksDetailScreen.kt
@@ -1,4 +1,4 @@
-package com.kirilpetriv.showcase.presentation
+package com.kirilpetriv.showcase.feature.artwork
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -26,8 +26,8 @@ import androidx.compose.ui.unit.dp
 import coil.compose.AsyncImage
 import com.kirilpetriv.showcase.R
 import com.kirilpetriv.showcase.models.Artwork
-import com.kirilpetriv.showcase.presentation.basecomposables.LoadingIndicatorScreen
-import com.kirilpetriv.showcase.presentation.basecomposables.ScreenErrorView
+import com.kirilpetriv.showcase.commonUi.LoadingIndicatorScreen
+import com.kirilpetriv.showcase.commonUi.ScreenErrorView
 import org.koin.androidx.compose.koinViewModel
 import org.koin.core.parameter.parametersOf
