+++ app/src/test/java/com/kirilpetriv/showcase/presentation/ArtworkViewModelTest.kt
@@ -1,6 +1,8 @@
 package com.kirilpetriv.showcase.presentation
 
 import com.kirilpetriv.showcase.core.ArtworkRepository
+import com.kirilpetriv.showcase.feature.artwork.ArtworkDetailScreenState
+import com.kirilpetriv.showcase.feature.artwork.ArtworksDetailViewModel
 import com.kirilpetriv.showcase.models.NetworkError
 import com.kirilpetriv.showcase.models.Resource
 import com.kirilpetriv.showcase.templates.artworkModelTemplate
